Refactored the battle queue to make things easier for future implementations
Refactored the battle queue to make things easier for future implementations
