+++ Assets/Scripts/Managers/BattleQueue.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
+using Unity.VisualScripting;
 using UnityEngine;
 using static UnityEngine.GraphicsBuffer;
 
@@ -21,7 +22,7 @@ public class BattleQueue : MonoBehaviour
         {
             BattleQueueInstance = this;
             playerActions = new List<ActionClass>();   
-            enemyActions = new List<ActionClass>(); 
+            enemyActions = new List<ActionClass>();
         }
         else if (BattleQueueInstance != this)
         {
@@ -45,64 +46,11 @@ public class BattleQueue : MonoBehaviour
         playerActions.Add(action);
         action.Origin = player;
 
-        ClashBothEntities(action.Origin, action.Target);
+        CardComparator.Instance.CompareCards(action, action);
     }
 
-    /*
-    EntityClass origin: Origin of the action card played
-    EntityClass target: Target of the action card played
+ 
 
-    Purpose: The two clashing enemies come together to clash, their positions will ideally be based off their speed
-    Then, whoever wins the clash should stagger the opponent backwards. 
-     */
-    private void ClashBothEntities(EntityClass origin, EntityClass target)
-    {
-        //The Distance weighting will be calculated based on speeds of the two clashing cards
-        Vector2 centeredDistance = (origin.myTransform.position * 0.3f + 0.7f * target.myTransform.position);
-        float bufferedRadius = 0.6f;
-        float duration = 0.6f;
-        StartCoroutine(origin.MoveToPosition(centeredDistance, bufferedRadius, duration));
-        StartCoroutine(target.MoveToPosition(centeredDistance, bufferedRadius, duration));
-
-        StartCoroutine(GeneralTimer(duration + 1f, StaggerEntities, new KeyValuePair<EntityClass, EntityClass>(origin, target)));
-    }
-
-    /*
-     * A Generic Timer function that performs a task when it is completed. 
-     * A delegate is defined here and is essentially a function pointer to a call back. Yippee !!
-     * 
-     * float duration: Duration until function is called
-     * TimerCompletedTask DoFinishedTask: A function pointer to a callback
-     * object parameter: generic paramter to pass in to the delegate
-     */
-    private delegate void TimerCompletedTask(object parameter);
-    private IEnumerator GeneralTimer(float duration, TimerCompletedTask DoFinishedTask, object parameter)
-    {
-        float elapsedTime = 0f;
-        while (elapsedTime < duration)
-        {
-            elapsedTime += Time.deltaTime;
-            yield return null;
-        }
-        Debug.Log("Timer is up, finished task is executing");
-        DoFinishedTask(parameter);
-    }
-
-    /*
-     Takes in a pair of entities with Key being the origin and value being the Target
-    Then it calculates a direction vector and staggers the Target bacl from the Origin.
-     */
-    private void StaggerEntities(object pair)
-    {
-        KeyValuePair<EntityClass, EntityClass> pairOfEntities = (KeyValuePair < EntityClass, EntityClass > ) pair;
-        Vector2 directionVector = pairOfEntities.Value.myTransform.position - pairOfEntities.Key.myTransform.position;
-
-        Vector2 normalizedDirection = directionVector.normalized;
-        float staggerPower = 2f; //Depending on percentage health lost
-
-        StartCoroutine(pairOfEntities.Value.StaggerBack(pairOfEntities.Value.myTransform.position + (Vector3) normalizedDirection * staggerPower));
-
-    }
 
     // FOR TESTING PURPOSES
     void AddRandomEnemyActions ()
